+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Change log
 -----------
 
+* Implement cache and other tiny fixes in resinhup wrappers [Andrei]
 * Backport dropbear atomic hostkey generation patch [Florin]
 * Update supervisor to v2.9.0 [Pablo]
 * Rewrite prepare-openvpn (style) and make it fail on error [Michal]